Extract open-close macro.
Pauan's make-w/close macro
(https://github.com/Pauan/ar/blob/old/arc.arc#L165) seemed like a good
idea to me, but I wondered if it was over-complicated for the need.
open-close is an experiment to see if using a second macro instead of
an expander helper function would serve as well.